Fix interrupt handling to comply with Visual6502
The changes in this commit makes each cycle equal to Visual6502 while
running the Klaus Dormann interrupt tests. The code is very hacky though
and need considerable clean up and restructuring.

Note that the interrupt tests does not seem to work on a real 6502
unless a small change is added, dealing with concurrent NMI and BRK
interrupts.